perf: add bundle splitting
- via lazy loading components and functions.
- compressing and chaning png->webp
- replace bundle heavy packages (lodash, react-icons)
