+++ frontend/react/index.html
@@ -7,11 +7,11 @@
     <title>FreshNest</title>
   </head>
   <body>
-    <link rel="preconnect" href="https://fresh-nest-flax.vercel.app" />.
+    <link rel="preconnect" href="https://fresh-nest-flax.vercel.app" />
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="dns-prefetch" href="https://accounts.google.com" />
 
-    <link rel="prefetch" href="/public/landing_bg.webp" as="image" />
+    <link rel="prefetch" href="/landing_bg.webp" as="image" />
 
     <div id="root"></div>
     <script type="module" src="/src/main.jsx"></script>
