+++ frontend/react/vite.config.js
@@ -5,6 +5,6 @@ import react from '@vitejs/plugin-react';
 export default defineConfig({
   plugins: [
     react()
-    // ,analyzer()
+    // , analyzer()
   ]
 });
